materials/activities/05-shiny-app/requirements.txt
shiny==0.4.0
ipywidgets==7.7.5
shinywidgets==0.2.1
ipyleaflet==0.17.3
loguru==0.7.0
shinyswatch==0.2.4
pandas==1.5.3
pyarrow==12.0.1
rich==13.4.2
ibis-framework[postgres]==5.1.0
vetiver==0.2.1
